changed directory cases
