+++ MicroservicesServerlessApp/Microservicess/scripts/deploy.py
